# internal/api/swagger/openapi.base.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "__TITLE__",
    "version": "1.0.0",
    "description": "Plataforma de gestão de atendimentos para psicologia. Toda resposta usa o envelope padrão { message, elapsed_ms, data, error, request_id }. Namespaces: /v1/admin (terapeuta, autenticado) e /v1/portal (paciente, sessão social)."
  },
  "servers": [
    { "url": "/", "description": "Instância local (TLS terminado no Pangolin)" }
  ],
  "components": {
...
    "securitySchemes": {
      "pangolinEmail": { "type": "apiKey", "in": "header", "name": "X-Pangolin-Email" },
      "pangolinSecret": { "type": "apiKey", "in": "header", "name": "X-Pangolin-Secret" },
      "portalSession": { "type": "apiKey", "in": "cookie", "name": "psicoman_portal" }
    }
  },
  "paths": {
...
    "/v1/admin/patients": {
      "get": { "tags": ["admin"], "summary": "Listar pacientes", "responses": { "200": { "description": "OK" } } },
      "post": { "tags": ["admin"], "summary": "Cadastrar paciente", "responses": { "201": { "description": "Criado" }, "409": { "description": "Email/CPF duplicado" }, "422": { "description": "Validação" } } }
    },
...
    "/v1/admin/patients/{id}/approve": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "post": { "tags": ["admin"], "summary": "Aprovar cadastro do paciente (libera o acesso ao portal)", "responses": { "200": { "description": "Aprovado" }, "404": { "description": "Não encontrado" }, "422": { "description": "Transição inválida" } } }
    },
    "/v1/admin/patients/{id}/reject": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "post": { "tags": ["admin"], "summary": "Rejeitar cadastro do paciente", "responses": { "200": { "description": "Rejeitado" }, "404": { "description": "Não encontrado" }, "422": { "description": "Transição inválida" } } }
    },
    "/v1/admin/patients/{id}": {
      "get": { "tags": ["admin"], "summary": "Obter paciente", "responses": { "200": { "description": "OK" }, "404": { "description": "Não encontrado" } } },
      "put": { "tags": ["admin"], "summary": "Atualizar paciente", "responses": { "200": { "description": "OK" } } },
      "delete": { "tags": ["admin"], "summary": "Remover paciente", "responses": { "200": { "description": "OK" } } }
    },
...
    "/v1/portal/login": { "post": { "tags": ["portal"], "summary": "Login social (credencial Google)", "responses": { "200": { "description": "OK" }, "401": { "description": "Falha" } } } },
    "/v1/portal/register": { "put": { "tags": ["portal"], "summary": "Cadastro básico (upsert por email)", "responses": { "200": { "description": "OK" }, "429": { "description": "Rate limit" } } } },
    "/v1/portal/approval-status": {
      "get": { "tags": ["portal"], "summary": "Estado de aprovação do meu cadastro (pendente/aprovado/rejeitado)", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "401": { "description": "Sem sessão" } } }
    },
    "/v1/portal/me": {
      "get": { "tags": ["portal"], "summary": "Meu perfil", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "401": { "description": "Sem sessão" } } },
      "put": { "tags": ["portal"], "summary": "Atualizar meu perfil (exige aprovação)", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "403": { "description": "Cadastro não aprovado" } } }
    },
    "/v1/portal/availability": { "get": { "tags": ["portal"], "summary": "Agenda aberta (exige aprovação)", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "403": { "description": "Cadastro não aprovado" } } } },
    "/v1/portal/appointment-requests": {
      "get": { "tags": ["portal"], "summary": "Meus pedidos (exige aprovação)", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "403": { "description": "Cadastro não aprovado" } } },
      "post": { "tags": ["portal"], "summary": "Solicitar sessão (exige aprovação)", "security": [{ "portalSession": [] }], "responses": { "201": { "description": "Criado" }, "403": { "description": "Cadastro não aprovado" }, "429": { "description": "Rate limit" } } }
    },
    "/v1/portal/sessions": { "get": { "tags": ["portal"], "summary": "Minhas sessões (exige aprovação)", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "403": { "description": "Cadastro não aprovado" } } } },
    "/v1/portal/debts": { "get": { "tags": ["portal"], "summary": "Meus débitos (exige aprovação)", "security": [{ "portalSession": [] }], "responses": { "200": { "description": "OK" }, "403": { "description": "Cadastro não aprovado" } } } }
  }
}